Moved notebook schema function to context
