+++ lib/cursif_web/resolvers/collaborators.ex
@@ -4,8 +4,6 @@ defmodule CursifWeb.Resolvers.Collaborators do
 
 	@spec add_collaborator(map(), map()) :: {:ok, Collaborator.t()}  | {:error, atom()}
 	def add_collaborator(collaborator, _context) do
-		Notebooks.get_notebook!(collaborator.notebook_id)
-			
 		case Notebooks.add_collaborator(collaborator) do
 			{:ok, collaborator} -> {:ok, collaborator}
 			{:error, changeset} -> {:error, changeset}
@@ -14,9 +12,7 @@ defmodule CursifWeb.Resolvers.Collaborators do
 
 	@spec delete_collaborator(map(), map()) :: {:ok, Collaborator.t()} | {:error, atom()}
 	def delete_collaborator(collaborator, _context) do
-		Notebooks.get_notebook!(collaborator.notebook_id)
-
-		case Notebooks.delete_collaborator_by_user_id(collaborator.notebook_id, collaborator.user_id) do
+		case Notebooks.delete_collaborator(collaborator) do
 			{1, nil} -> {:ok, %{message: "collaborator removed successfully"}}
 			{0, nil} -> {:error, :not_found}
 		end
