+++ lib/cursif_web/resolvers/notebooks.ex
@@ -42,5 +42,5 @@ defmodule CursifWeb.Resolvers.Notebooks do
 
   @spec get_owner(map(), map(), map()) :: {:ok, User.t()}
   def get_owner(notebook, _args, _context),
-    do: {:ok, Notebook.owner(notebook)}
+    do: {:ok, Notebooks.get_owner(notebook)}
 end
